+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =======
 
 * Add wizzard widget.
+* Add basket widget.
 
 0.1.0.0
 =======